requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
aiosqlite==0.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
docker==7.1.0
cryptography==46.0.5
qrcode[pil]==7.4.2
bcrypt==4.1.3
aiofiles==23.2.1
sse-starlette==2.1.3
python-multipart==0.0.9
pyyaml==6.0.2
python-dotenv==1.0.1
slowapi==0.1.9
pywebpush==2.0.0
psutil==6.0.0
numpy>=1.22.2
pyaudio==0.2.14
python-vlc>=3.0.20115
websockets>=12.0
num2words==0.5.14
piper-tts>=1.4.0
# transliterate removed — replaced by split_by_language() + dual PiperVoice

# Pinned transitive dependencies (security)
h11>=0.16.0
anyio>=4.4.0
starlette>=0.40.0,<0.41.0
zipp>=3.19.1
